Tab Trainers djelimicno uradje i funkcionalno dugme delete Trainer

+++ src/ba/unsa/etf/rs/projekat/MainController.java
@@ -31,6 +31,14 @@ public class MainController implements Initializable {
     public ImageView pictureRider;
     public TextField surnameRider;
     public TextField nameRider;
+    public ListView<Trainer> trainerList;
+    public TextField nameTrainer;
+    public TextField surnameTrainer;
+    public ImageView pictureTrainer;
+    public TextField jmbgTrainer;
+    public TextField birthTrainer;
+    public TextField membershipTrainer;
+    public TextField salaryTrainer;
     private RidingDAO dao =new RidingDAOBase();
 
 
@@ -71,7 +79,27 @@ public class MainController implements Initializable {
                 jmbgRider.setText(newKorisnik.getJmbg());
                 pictureRider.setImage(new Image(newKorisnik.getPicture()));
             }
-
+        });
+        //tab Trainer
+        trainerList.setItems(dao.getTrainers());
+        trainerList.getSelectionModel().selectedItemProperty().addListener((obs, oldKorisnik, newKorisnik) -> {
+            if(newKorisnik==null){
+                nameTrainer.setText("");
+                surnameTrainer.setText("");
+                birthTrainer.setText("");
+                membershipTrainer.setText("");
+                jmbgTrainer.setText("");
+                pictureTrainer.setImage(null);
+                salaryTrainer.setText("");
+            }else{
+                nameTrainer.setText(newKorisnik.getRider().getName());
+                surnameTrainer.setText(newKorisnik.getRider().getSurname());
+                birthTrainer.setText(newKorisnik.getRider().getDateOfBirth().toString());
+                membershipTrainer.setText(newKorisnik.getRider().getDateOfMembership().toString());
+                jmbgTrainer.setText(newKorisnik.getRider().getJmbg());
+                pictureTrainer.setImage(new Image(newKorisnik.getRider().getPicture()));
+                salaryTrainer.setText(String.valueOf(newKorisnik.getSalary()));
+            }
         });
     }
 
@@ -83,20 +111,53 @@ public class MainController implements Initializable {
             confirm.setTitle("Delete");
             Optional<ButtonType> result = confirm.showAndWait();
             if (result.get() == ButtonType.OK) {
-                dao.deleteRider(r);
-                confirm.close();
-            } else {
-                confirm.close();
-
+                int j = 0;
+                for(int i =0;i<dao.getTrainers().size();i++){
+                    if(dao.getTrainers().get(i).getRider().getId() == r.getId()){
+                        j++;
+                    }
+                }
+                if(j == 0){
+                    dao.deleteRider(r);
+                }else{
+                    Alert error = new Alert(Alert.AlertType.ERROR);
+                    error.setContentText("You cannot delete rider who is a trainer");
+                    error.setTitle("Deleteing failed");
+                    error.show();
+                }
             }
+            confirm.close();
             ridersList.getSelectionModel().selectFirst();
             ridersList.setItems(dao.getRiders());
+
         }else{
             Alert error = new Alert(Alert.AlertType.ERROR);
             error.setContentText("No rider was selected!");
-            error.setTitle("Deleteing feiled");
+            error.setTitle("Deleteing failed");
             error.show();
         }
 
+    }
+
+    public void trainerDelete(ActionEvent actionEvent) {
+        Trainer t =trainerList.getSelectionModel().getSelectedItem();
+        if(t!=null) {
+            Alert confirm = new Alert(Alert.AlertType.CONFIRMATION);
+            confirm.setContentText("Are you sure you want to delete selected trainer?");
+            confirm.setTitle("Delete");
+            Optional<ButtonType> result = confirm.showAndWait();
+            if (result.get() == ButtonType.OK) {
+                dao.deleteTrainer(t);
+            }
+            confirm.close();
+            trainerList.getSelectionModel().selectFirst();
+            trainerList.setItems(dao.getTrainers());
+
+        }else{
+            Alert error = new Alert(Alert.AlertType.ERROR);
+            error.setContentText("No trainer was selected!");
+            error.setTitle("Deleteing failed");
+            error.show();
+        }
     }
 }
